Use new, but pretty much the same, tETH faucet
